Remove legacy election packages (#1426)
* Remove legacy election packages
* Retain a useful test from the legacy package
