+++ util/election2/etcd/election_test.go
@@ -15,12 +15,47 @@
 package etcd
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
 	"github.com/google/trillian/testonly/integration/etcd"
 	"github.com/google/trillian/util/election2/testonly"
 )
+
+func TestElectionThroughCommonClient(t *testing.T) {
+	_, client, cleanup, err := etcd.StartEtcd()
+	if err != nil {
+		t.Fatalf("StartEtcd(): %v", err)
+	}
+	defer cleanup()
+
+	ctx := context.Background()
+	fact := NewFactory("serv", client, "res/")
+
+	el1, err := fact.NewElection(ctx, "10")
+	if err != nil {
+		t.Fatalf("NewElection(10): %v", err)
+	}
+	el2, err := fact.NewElection(ctx, "20")
+	if err != nil {
+		t.Fatalf("NewElection(20): %v", err)
+	}
+
+	if err := el1.Await(ctx); err != nil {
+		t.Fatalf("Await(10): %v", err)
+	}
+	if err := el2.Await(ctx); err != nil {
+		t.Fatalf("Await(20): %v", err)
+	}
+
+	if err := el1.Close(ctx); err != nil {
+		t.Fatalf("Close(10): %v", err)
+	}
+	if err := el2.Close(ctx); err != nil {
+		t.Fatalf("Close(20): %v", err)
+	}
+}
 
 func TestElection(t *testing.T) {
 	_, client, cleanup, err := etcd.StartEtcd()
